server/middleware_test.go
package server

import (
	"context"
	"testing"
	"time"

	"go.flipt.io/flipt/config"
	"go.flipt.io/flipt/errors"
	flipt "go.flipt.io/flipt/rpc/flipt"
	"go.flipt.io/flipt/server/cache/memory"
	"go.flipt.io/flipt/storage"

	"github.com/sirupsen/logrus/hooks/test"
	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/mock"
	"github.com/stretchr/testify/require"
	"google.golang.org/grpc"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/status"
)

var logger, _ = test.NewNullLogger()

type validatable struct {
	err error
}

func (v *validatable) Validate() error {
	return v.err
}

func TestValidationUnaryInterceptor(t *testing.T) {
	tests := []struct {
		name       string
		req        interface{}
		wantCalled int
	}{
		{
			name:       "does not implement Validate",
			req:        struct{}{},
			wantCalled: 1,
		},
		{
			name:       "implements validate no error",
			req:        &validatable{},
			wantCalled: 1,
		},
		{
			name: "implements validate error",
			req:  &validatable{err: errors.New("invalid")},
		},
	}

	for _, tt := range tests {
		var (
			req        = tt.req
			wantCalled = tt.wantCalled
			called     int
		)

		t.Run(tt.name, func(t *testing.T) {
			var (
				spyHandler = grpc.UnaryHandler(func(ctx context.Context, req interface{}) (interface{}, error) {
					called++
					return nil, nil
				})
			)

			_, _ = ValidationUnaryInterceptor(context.Background(), req, nil, spyHandler)
			assert.Equal(t, wantCalled, called)
		})
	}
}

func TestErrorUnaryInterceptor(t *testing.T) {
	tests := []struct {
		name     string
		wantErr  error
		wantCode codes.Code
	}{
		{
			name:     "not found error",
			wantErr:  errors.ErrNotFound("foo"),
			wantCode: codes.NotFound,
		},
		{
			name:     "invalid error",
			wantErr:  errors.ErrInvalid("foo"),
			wantCode: codes.InvalidArgument,
		},
		{
			name:     "invalid field",
			wantErr:  errors.InvalidFieldError("bar", "is wrong"),
			wantCode: codes.InvalidArgument,
		},
		{
			name:     "empty field",
			wantErr:  errors.EmptyFieldError("bar"),
			wantCode: codes.InvalidArgument,
		},
		{
			name:     "other error",
			wantErr:  errors.New("foo"),
			wantCode: codes.Internal,
		},
		{
			name: "no error",
		},
	}

	for _, tt := range tests {
		var (
			wantErr  = tt.wantErr
			wantCode = tt.wantCode
		)

		t.Run(tt.name, func(t *testing.T) {
			var (
				spyHandler = grpc.UnaryHandler(func(ctx context.Context, req interface{}) (interface{}, error) {
					return nil, wantErr
				})
			)

			_, err := ErrorUnaryInterceptor(context.Background(), nil, nil, spyHandler)
			if wantErr != nil {
				require.Error(t, err)
				status := status.Convert(err)
				assert.Equal(t, wantCode, status.Code())
				return
			}

			require.NoError(t, err)
		})
	}
}

func TestCacheUnaryInterceptor_GetFlag(t *testing.T) {
	var (
		store = &storeMock{}
		cache = memory.NewCache(config.CacheConfig{
			TTL:     time.Second,
			Enabled: true,
			Backend: config.CacheMemory,
		})
		cacheSpy = newCacheSpy(cache)
		s        = &Server{
			logger: logger,
			store:  store,
		}
		req = &flipt.GetFlagRequest{Key: "foo"}
	)

	store.On("GetFlag", mock.Anything, "foo").Return(&flipt.Flag{
		Key:     req.Key,
		Enabled: true,
	}, nil)

	unaryInterceptor := CacheUnaryInterceptor(cacheSpy, logger)

	handler := func(ctx context.Context, r interface{}) (interface{}, error) {
		return s.GetFlag(ctx, r.(*flipt.GetFlagRequest))
	}

	info := &grpc.UnaryServerInfo{
		FullMethod: "FakeMethod",
	}

	for i := 0; i < 10; i++ {
		got, err := unaryInterceptor(context.Background(), req, info, handler)
		require.NoError(t, err)
		assert.NotNil(t, got)
	}

	assert.Equal(t, 10, cacheSpy.getCalled)
	assert.NotEmpty(t, cacheSpy.getKeys)

	// cache key is flipt:(md5(f:foo))
	const cacheKey = "flipt:864ce319cc64891a59e4745fbe7ecc47"
	_, ok := cacheSpy.getKeys[cacheKey]
	assert.True(t, ok)

	assert.Equal(t, 1, cacheSpy.setCalled)
	assert.NotEmpty(t, cacheSpy.setItems)
	assert.NotEmpty(t, cacheSpy.setItems[cacheKey])
}

func TestCacheUnaryInterceptor_UpdateFlag(t *testing.T) {
	var (
		store = &storeMock{}
		cache = memory.NewCache(config.CacheConfig{
			TTL:     time.Second,
			Enabled: true,
			Backend: config.CacheMemory,
		})
		cacheSpy = newCacheSpy(cache)
		s        = &Server{
			logger: logger,
			store:  store,
		}
		req = &flipt.UpdateFlagRequest{
			Key:         "key",
			Name:        "name",
			Description: "desc",
			Enabled:     true,
		}
	)

	store.On("UpdateFlag", mock.Anything, req).Return(&flipt.Flag{
		Key:         req.Key,
		Name:        req.Name,
		Description: req.Description,
		Enabled:     req.Enabled,
	}, nil)

	unaryInterceptor := CacheUnaryInterceptor(cacheSpy, logger)

	handler := func(ctx context.Context, r interface{}) (interface{}, error) {
		return s.UpdateFlag(ctx, r.(*flipt.UpdateFlagRequest))
	}

	info := &grpc.UnaryServerInfo{
		FullMethod: "FakeMethod",
	}

	got, err := unaryInterceptor(context.Background(), req, info, handler)
	require.NoError(t, err)
	assert.NotNil(t, got)

	assert.Equal(t, 1, cacheSpy.deleteCalled)
	assert.NotEmpty(t, cacheSpy.deleteKeys)
}

func TestCacheUnaryInterceptor_DeleteFlag(t *testing.T) {
	var (
		store = &storeMock{}
		cache = memory.NewCache(config.CacheConfig{
			TTL:     time.Second,
			Enabled: true,
			Backend: config.CacheMemory,
		})
		cacheSpy = newCacheSpy(cache)
		s        = &Server{
			logger: logger,
			store:  store,
		}
		req = &flipt.DeleteFlagRequest{
			Key: "key",
		}
	)

	store.On("DeleteFlag", mock.Anything, req).Return(nil)

	unaryInterceptor := CacheUnaryInterceptor(cacheSpy, logger)

	handler := func(ctx context.Context, r interface{}) (interface{}, error) {
		return s.DeleteFlag(ctx, r.(*flipt.DeleteFlagRequest))
	}

	info := &grpc.UnaryServerInfo{
		FullMethod: "FakeMethod",
	}

	got, err := unaryInterceptor(context.Background(), req, info, handler)
	require.NoError(t, err)
	assert.NotNil(t, got)

	assert.Equal(t, 1, cacheSpy.deleteCalled)
	assert.NotEmpty(t, cacheSpy.deleteKeys)
}

func TestCacheUnaryInterceptor_CreateVariant(t *testing.T) {
	var (
		store = &storeMock{}
		cache = memory.NewCache(config.CacheConfig{
			TTL:     time.Second,
			Enabled: true,
			Backend: config.CacheMemory,
		})
		cacheSpy = newCacheSpy(cache)
		s        = &Server{
			logger: logger,
			store:  store,
		}
		req = &flipt.CreateVariantRequest{
			FlagKey:     "flagKey",
			Key:         "key",
			Name:        "name",
			Description: "desc",
		}
	)

	store.On("CreateVariant", mock.Anything, req).Return(&flipt.Variant{
		Id:          "1",
		FlagKey:     req.FlagKey,
		Key:         req.Key,
		Name:        req.Name,
		Description: req.Description,
		Attachment:  req.Attachment,
	}, nil)

	unaryInterceptor := CacheUnaryInterceptor(cacheSpy, logger)

	handler := func(ctx context.Context, r interface{}) (interface{}, error) {
		return s.CreateVariant(ctx, r.(*flipt.CreateVariantRequest))
	}

	info := &grpc.UnaryServerInfo{
		FullMethod: "FakeMethod",
	}

	got, err := unaryInterceptor(context.Background(), req, info, handler)
	require.NoError(t, err)
	assert.NotNil(t, got)

	assert.Equal(t, 1, cacheSpy.deleteCalled)
	assert.NotEmpty(t, cacheSpy.deleteKeys)
}

func TestCacheUnaryInterceptor_UpdateVariant(t *testing.T) {
	var (
		store = &storeMock{}
		cache = memory.NewCache(config.CacheConfig{
			TTL:     time.Second,
			Enabled: true,
			Backend: config.CacheMemory,
		})
		cacheSpy = newCacheSpy(cache)
		s        = &Server{
			logger: logger,
			store:  store,
		}
		req = &flipt.UpdateVariantRequest{
			Id:          "1",
			FlagKey:     "flagKey",
			Key:         "key",
			Name:        "name",
			Description: "desc",
		}
	)

	store.On("UpdateVariant", mock.Anything, req).Return(&flipt.Variant{
		Id:          req.Id,
		FlagKey:     req.FlagKey,
		Key:         req.Key,
		Name:        req.Name,
		Description: req.Description,
		Attachment:  req.Attachment,
	}, nil)

	unaryInterceptor := CacheUnaryInterceptor(cacheSpy, logger)

	handler := func(ctx context.Context, r interface{}) (interface{}, error) {
		return s.UpdateVariant(ctx, r.(*flipt.UpdateVariantRequest))
	}

	info := &grpc.UnaryServerInfo{
		FullMethod: "FakeMethod",
	}

	got, err := unaryInterceptor(context.Background(), req, info, handler)
	require.NoError(t, err)
	assert.NotNil(t, got)

	assert.Equal(t, 1, cacheSpy.deleteCalled)
	assert.NotEmpty(t, cacheSpy.deleteKeys)
}

func TestCacheUnaryInterceptor_DeleteVariant(t *testing.T) {
	var (
		store = &storeMock{}
		cache = memory.NewCache(config.CacheConfig{
			TTL:     time.Second,
			Enabled: true,
			Backend: config.CacheMemory,
		})
		cacheSpy = newCacheSpy(cache)
		s        = &Server{
			logger: logger,
			store:  store,
		}
		req = &flipt.DeleteVariantRequest{
			Id: "1",
		}
	)

	store.On("DeleteVariant", mock.Anything, req).Return(nil)

	unaryInterceptor := CacheUnaryInterceptor(cacheSpy, logger)

	handler := func(ctx context.Context, r interface{}) (interface{}, error) {
		return s.DeleteVariant(ctx, r.(*flipt.DeleteVariantRequest))
	}

	info := &grpc.UnaryServerInfo{
		FullMethod: "FakeMethod",
	}

	got, err := unaryInterceptor(context.Background(), req, info, handler)
	require.NoError(t, err)
	assert.NotNil(t, got)

	assert.Equal(t, 1, cacheSpy.deleteCalled)
	assert.NotEmpty(t, cacheSpy.deleteKeys)
}

func TestCacheUnaryInterceptor_Evaluate(t *testing.T) {
	var (
		store = &storeMock{}
		cache = memory.NewCache(config.CacheConfig{
			TTL:     time.Second,
			Enabled: true,
			Backend: config.CacheMemory,
		})
		cacheSpy = newCacheSpy(cache)
		s        = &Server{
			logger: logger,
			store:  store,
		}
	)

	store.On("GetFlag", mock.Anything, "foo").Return(enabledFlag, nil)

	store.On("GetEvaluationRules", mock.Anything, "foo").Return(
		[]*storage.EvaluationRule{
			{
				ID:               "1",
				FlagKey:          "foo",
				SegmentKey:       "bar",
				SegmentMatchType: flipt.MatchType_ALL_MATCH_TYPE,
				Rank:             0,
				Constraints: []storage.EvaluationConstraint{
					// constraint: bar (string) == baz
					{
						ID:       "2",
						Type:     flipt.ComparisonType_STRING_COMPARISON_TYPE,
						Property: "bar",
						Operator: flipt.OpEQ,
						Value:    "baz",
					},
					// constraint: admin (bool) == true
					{
						ID:       "3",
						Type:     flipt.ComparisonType_BOOLEAN_COMPARISON_TYPE,
						Property: "admin",
						Operator: flipt.OpTrue,
					},
				},
			},
		}, nil)

	store.On("GetEvaluationDistributions", mock.Anything, "1").Return(
		[]*storage.EvaluationDistribution{
			{
				ID:                "4",
				RuleID:            "1",
				VariantID:         "5",
				Rollout:           100,
				VariantKey:        "boz",
				VariantAttachment: `{"key":"value"}`,
			},
		}, nil)

	tests := []struct {
		name      string
		req       *flipt.EvaluationRequest
		wantMatch bool
	}{
		{
			name: "matches all",
			req: &flipt.EvaluationRequest{
				FlagKey:  "foo",
				EntityId: "1",
				Context: map[string]string{
					"bar":   "baz",
					"admin": "true",
				},
			},
			wantMatch: true,
		},
		{
			name: "no match all",
			req: &flipt.EvaluationRequest{
				FlagKey:  "foo",
				EntityId: "1",
				Context: map[string]string{
					"bar":   "boz",
					"admin": "true",
				},
			},
		},
		{
			name: "no match just bool value",
			req: &flipt.EvaluationRequest{
				FlagKey:  "foo",
				EntityId: "1",
				Context: map[string]string{
					"admin": "true",
				},
			},
		},
		{
			name: "no match just string value",
			req: &flipt.EvaluationRequest{
				FlagKey:  "foo",
				EntityId: "1",
				Context: map[string]string{
					"bar": "baz",
				},
			},
		},
	}

	unaryInterceptor := CacheUnaryInterceptor(cacheSpy, logger)

	handler := func(ctx context.Context, r interface{}) (interface{}, error) {
		return s.Evaluate(ctx, r.(*flipt.EvaluationRequest))
	}

	info := &grpc.UnaryServerInfo{
		FullMethod: "FakeMethod",
	}

	for i, tt := range tests {
		var (
			i         = i + 1
			req       = tt.req
			wantMatch = tt.wantMatch
		)

		t.Run(tt.name, func(t *testing.T) {
			got, err := unaryInterceptor(context.Background(), req, info, handler)
			require.NoError(t, err)
			assert.NotNil(t, got)

			resp := got.(*flipt.EvaluationResponse)
			assert.NotNil(t, resp)
			assert.Equal(t, "foo", resp.FlagKey)
			assert.Equal(t, req.Context, resp.RequestContext)

			assert.Equal(t, i, cacheSpy.getCalled)
			assert.NotEmpty(t, cacheSpy.getKeys)

			if !wantMatch {
				assert.False(t, resp.Match)
				assert.Empty(t, resp.SegmentKey)
				return
			}

			assert.True(t, resp.Match)
			assert.Equal(t, "bar", resp.SegmentKey)
			assert.Equal(t, "boz", resp.Value)
			assert.Equal(t, `{"key":"value"}`, resp.Attachment)
		})
	}
}
